Avoid compilation warnings

+++ EventFiltering/PWGUD/diffractionFilter.cxx
@@ -75,7 +75,7 @@ T compatibleBCs(soa::Join<aod::Collisions, aod::EvSels>::iterator const& collisi
 
   bcIter.moveByIndex(-moveCount); // Move back to original position
   int64_t minBCId = collision.bcId();
-  while (bcIter != bcs.begin() && bcIter.globalBC() <= maxBC && bcIter.globalBC() >= minBC) {
+  while (bcIter != bcs.begin() && bcIter.globalBC() <= maxBC && (int64_t)bcIter.globalBC() >= minBC) {
     LOGF(debug, "Table id %d BC %llu", bcIter.globalIndex(), bcIter.globalBC());
     minBCId = bcIter.globalIndex();
     --bcIter;
